Split Backend.Specialise into smaller modules
